+++ localstack/utils/common.py
@@ -172,6 +172,9 @@ from localstack.utils.time import (  # noqa
     timestamp_millis,
 )
 
+# TODO: remove imports from here (need to update any client code that imports these from utils.common)
+from localstack.utils.xml import obj_to_xml, strip_xmlns  # noqa
+
 # set up logger
 LOG = logging.getLogger(__name__)
 
@@ -636,33 +639,6 @@ def edge_ports_info():
 
 def base64_to_hex(b64_string: str) -> bytes:
     return binascii.hexlify(base64.b64decode(b64_string))
-
-
-def obj_to_xml(obj: Any) -> str:
-    """Return an XML representation of the given object (dict, list, or primitive).
-    Does NOT add a common root element if the given obj is a list.
-    Does NOT work for nested dict structures."""
-    if isinstance(obj, list):
-        return "".join([obj_to_xml(o) for o in obj])
-    if isinstance(obj, dict):
-        return "".join(["<{k}>{v}</{k}>".format(k=k, v=obj_to_xml(v)) for (k, v) in obj.items()])
-    return str(obj)
-
-
-def strip_xmlns(obj: Any) -> Any:
-    """Strip xmlns attributes from a dict returned by xmltodict.parse."""
-    if isinstance(obj, list):
-        return [strip_xmlns(item) for item in obj]
-    if isinstance(obj, dict):
-        # Remove xmlns attribute.
-        obj.pop("@xmlns", None)
-        if len(obj) == 1 and "#text" in obj:
-            # If the only remaining key is the #text key, elide the dict
-            # entirely, to match the structure that xmltodict.parse would have
-            # returned if the xmlns namespace hadn't been present.
-            return obj["#text"]
-        return {k: strip_xmlns(v) for k, v in obj.items()}
-    return obj
 
 
 def get_proxies() -> Dict[str, str]:
